+++ src/server/templates/helpers/page_title.ts
@@ -1,0 +1,5 @@
+export interface PageTitleDefaults {
+  prefix?: string;
+  suffix?: string;
+  title: string;
+}
